\name{crimpsoftware-package}
\alias{crimpsoftware-package}
\alias{crimpsoftware}
\title{
Calculations Related to the Documents 'crimp.pdf' and 'crimpmeas.pdf'
}
\description{
Given amplidude and wavelength and staple length calculates fibre l
ength and fibre length to staple length ratio for unfolded and stretched helix c
rimp types. For unfolded helix also calculates angle between unfoldings.
}
\details{
This package provides functions for calculation of fibre length and fibre length to staple length ratio, given crimp amplitude and wavelength and staple length.
}
\author{
Neville Jackson
}
\references{
  crimp.pdf, crimpmeas.pdf available in \url{https://github.com/nevillejackson/Fleece-biology/crimpformation}
}
\keyword{ package }
\seealso{
Functions pred() and predlf()
}
\examples{
   
}
