\name{curvtoirad}
\alias{curvtoirad}
\title{
Curvature in degrees per mm to intrinsic radius of curvature in mm
}
\description{
Curvature in degrees per mm to intrinsic radius of curvature in mm
}
\usage{
curvtoirad(curv)
}
\arguments{
  \item{curv}{
curvature in degrees per mm
}
}
\details{
Fibre curvature is defined as the number of degrees ( or radians) subtended by a 1mm long fibre snippet at the centre of curvature
}
\value{
Returns intrinsic radius of curvature in mm
}
\references{
}
\author{
Neville Jackson
}
\note{
}


\seealso{
}
\examples{
}
\keyword{misc}
