\name{degtorad}
\alias{degtorad}
\title{
Converts angles in degrees to radians
}
\description{
Converts angles in degrees to radians
}
\usage{
degtorad(x)
}
\arguments{
  \item{x}{
Vector of angles in degrees
}
}
\details{

}
\value{
Vector of angles in radians
}
\references{
}
\author{
Neville Jackson
}

\seealso{
Function radtodeg()
}
\examples{
degtorad(45)
}
\keyword{ misc }
