\name{iradtocurv}
\alias{iradtocurv}
\title{
Convert intrinsic radius to fibre curvature
}
\description{
Convert intrinsic radius in mm to fibre curvature in both rads per mm  and degrees per mm
}
\usage{
iradtocurv(irad)
}
\arguments{
  \item{irad}{
intrinsic radius of curvature in mm
}
}
\details{
Curvature in radians per mm is the reciprocal of intrinsic radius of curvature in mm. The function also converts rads per mm to degrees per mm.
}
\value{
\item{curv[1]}{curvature in radians per mm}
\item{curv[2]}{curvature in degrees per mm}
}
\references{
}
\author{
Neville Jackson
}
\note{
}


\seealso{

}
\examples{
}
\keyword{ misc }
