\name{radtodeg}
\alias{radtodeg}
\title{
Converts angles in radians to degrees
}
\description{
Converts angles in radians to degrees
}
\usage{
radtodeg(x)
}
\arguments{
  \item{x}{
Vector of angles in radians
}
}
\details{

}
\value{
Vector of angles in degrees
}
\references{
}
\author{
Neville Jackson
}
\note{
}

\seealso{
Function degtorad()
}
\examples{
}
\keyword{ misc }
