\name{crimpsoftware-package}
\alias{crimpsoftware-package}
\alias{crimpsoftware}
\docType{package}
\title{
\packageTitle{crimpsoftware}
}
\description{
\packageDescription{crimpsoftware}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{crimpsoftware}
\packageIndices{crimpsoftware}
This package provides functions for calculation of fibre length and fibre length to staple length ration, given crimp amplitude and wavelength and staple length.
}
\author{
\packageAuthor{crimpsoftware}

Maintainer: \packageMaintainer{crimpsoftware}
}
\references{
  crimp.pdf, crimpmeas.pdf available in \url{https://github.com/nevillejackson/Fleece-biology/crimpformation}
}
\keyword{ package }
\seealso{
Functions pred() and predlf()
}
\examples{
   
}
